refactor(core): DefaultResolver 삭제, lib 생성
- lib에는 외부에서 사용할 API가 담기게 됨
