--- packages/core/src/decl/features/default.ts
@@ -1,4 +1,0 @@
-export interface DefaultResolverContext {
-  // Dialog 자신을 삭제합니다.
-  remove: () => void;
-}
