--- packages/core/src/features/default/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useDefaultResolver } from "@/features/default/hooks/external/useDefaultResolver";
