+++ packages/core/src/index.ts
@@ -5,20 +5,13 @@ export { default as DialogArea } from "@/provider/components/DialogArea";
 /* resolver */
 export { default as DialogResolver } from "@/resolver/components/DialogResolver";
 
-/* default resolver */
-export * from "@/features/default/hooks";
-
 /* patch resolver */
 export * from "@/features/patch/hooks";
 
 /* event resolver */
 export * from "@/features/event/hooks";
 
-/* resolver utils */
-export * from "@/utils/resolver";
-
-/* patch utils */
-export * from "@/utils/patch";
-
-/* type declarations */
-export type { DialogID, DialogKey, Dialog } from "$";
+/* creators */
+export * from "@/lib/creator/resolver";
+export * from "@/lib/creator/patch";
+export * from "@/lib/creator/dialog";
