+++ packages/core/src/lib/creator/dialog.ts
@@ -1,0 +1,6 @@
+import { Dialog, DialogComponentProfile } from "$";
+
+export const createDialog = <T extends object>(
+  dialog: Dialog,
+  template?: DialogComponentProfile<T>
+) => {};
